templates/boilerplate/server-python/requirements.txt
aiohttp
python-socketio
#sanic
flask
#tornado
eventlet